Crate sgx_tkey_exchange

Trusted Key Exchange Library

The library allow an ISV to exchange secrets between its server and its enclaves. They are used in concert with untrusted Key Exchange functions.

Functions

rsgx_ra_close
rsgx_ra_close release context created by rsgx_ra_init or rsgx_ra_init_ex.

rsgx_ra_get_keys
The sgx_ra_get_keys function is used to get the negotiated keys of a remote attestation and key exchange session.

rsgx_ra_init
The rsgx_ra_init function creates a context for the remote attestation and key exchange process.

rsgx_ra_init_ex
The rsgx_ra_init_ex function creates a context for the remote attestation and key exchange process while it allows the use of a custom defined Key Derivation Function (KDF).