Function sgx_tkey_exchange::rsgx_ra_close

pub fn rsgx_ra_close(context: sgx_ra_context_t) -> SgxError

rsgx_ra_close release context created by rsgx_ra_init or rsgx_ra_init_ex.

Call the rsgx_ra_close function to release the remote attestation and key exchange context after the process is done and the context isn’t needed anymore.

Description

At the end of a key exchange process, the caller needs to use this API in an enclave to clear and free memory associated with this remote attestation session.

Parameters

context

Context returned by rsgx_ra_init.

Requirements

Header: sgx_tkey_exchange.edl

Library: libsgx_tkey_exchange.a

Errors

SGX_ERROR_INVALID_PARAMETER

Indicates an error that the input parameters are invalid.